Implemented Drawer class
The Drawer class simplifies the task of drawing lines and rectangles to
the screen, whether they be for debugging purposes or for visual flair.
Just declare one in whatever class you're working in and pass the
current game-state (which the screen should be receiving in its
constructor anyway) to set it up.