Ui dev (#112)
* Tooltip now appears on-screen after hovering over a tile for 0.5 seconds

* Fixed further memory leaks and adjusted tooltip such that it never ventures outside of the game's window-space

Sorry for taking so long with what seems like a relatively minor commit:
it's only so minor because I ended up scrapping over two hours of work
due to a plethora of issues that I don't have the time or the energy to
explain right now. Proper documentation will be coming Soon™.

* Updated GameScreen UI (WIP)

Started to add the rest of the main elements to the actual GameScreen,
although it isn't quite done yet. Roboticon upgrade panel, market and
pause button need to be added.

* Further Changes to GameScreen UI

* Revert "Further Changes to GameScreen UI"

This reverts commit 97d5f61e756df8cab4411a2b162bc944c7e1df26.

* More UI updates

Coming along nicely!

* Saving recent UI changes in preparation for workstation transition

* Improved menu; added new fonts; added pausing functionality; added college symbols

Getting close to finishing the interface now: just need to add volume
sliders, a market interface, new font(s) for the game-screen, a tutorial
screen and a placeholder leaderboard

* Added two more college symbols

For some reason, these weren't pushed last time